Renamed to head->heads in image/classification
